requirements.txt
wikibaseintegrator=0.11.1
python-dotenv=0.19.2
